refactor(desktop): extract top-level utilities from layout.ts (Step 1)
Extract 4 top-level utility functions from desktop/layout.ts into focused modules:
- getCurrentColors() → utils/colorUtils.ts
- renderRule() → utils/ruleRenderer.ts
- updateStatisticsDisplay() → utils/statsUpdater.ts
- handleCanvasClick() → utils/canvasInteraction.ts

Benefits:
- Reduced layout.ts from 1613 to 1372 lines (-241 lines)
- Improved code organization and modularity
- Functions can now be tested in isolation
- Clearer separation of concerns

Part of Step 1 of the refactoring plan outlined in issue #148.
Subsequent steps will extract internal functions and event handlers.
